Fixed a format specification mismatch.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17995 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/snmp-notification-mib/snmpNotifyFilterTable/snmpNotifyFilterTable_data_access.c
@@ -243,7 +243,8 @@ snmpNotifyFilterTable_container_load(netsnmp_container *container)
         ++count;
     }
 
-    DEBUGMSGT(("verbose:snmpNotifyFilterTable:snmpNotifyFilterTable_container_load", "inserted %lu records\n", count));
+    DEBUGMSGT(("verbose:snmpNotifyFilterTable:snmpNotifyFilterTable_container_load",
+               "inserted %lu records\n", (long)count));
 
     return MFD_SUCCESS;
 }                               /* snmpNotifyFilterTable_container_load */
